Move deprecated functions to compat
