+++ glue/core/data_factories/hdf5.py
@@ -30,50 +30,6 @@ def extract_hdf5_datasets(handle):
     return datasets
 
 
-def extract_data_hdf5(filename, use_datasets='all'):
-    '''
-    Extract non-tabular datasets from an HDF5 file. If `use_datasets` is
-    'all', then all non-tabular datasets are extracted, otherwise only the
-    ones specified by `use_datasets` are extracted (`use_datasets` should
-    then contain a list of paths). If the requested datasets do not have
-    the same dimensions, an Exception is raised.
-    '''
-
-    import h5py
-
-    # Open file
-    file_handle = h5py.File(filename, 'r')
-
-    # Define function to read
-
-    # Read in all datasets
-    datasets = extract_hdf5_datasets(file_handle)
-
-    # Only keep non-tabular datasets
-    remove = []
-    for key in datasets:
-        if datasets[key].dtype.fields is not None:
-            remove.append(key)
-    for key in remove:
-        datasets.pop(key)
-
-    # Check that dimensions of all datasets are the same
-    reference_shape = datasets[list(datasets.keys())[0]].value.shape
-    for key in datasets:
-        if datasets[key].value.shape != reference_shape:
-            raise Exception("Datasets are not all the same dimensions")
-
-    # Extract data
-    arrays = {}
-    for key in datasets:
-        arrays[key] = datasets[key].value
-
-    # Close HDF5 file
-    file_handle.close()
-
-    return arrays
-
-
 def is_hdf5(filename):
     # All hdf5 files begin with the same sequence
     with open(filename, 'rb') as infile:
